QRangeModel: Change int by qreals
We want a better fine tunning of our values in order to make a better use
of the model (for example on progressbars or while animating values).

+++ src/models/qrangemodel.h
@@ -36,14 +36,14 @@ class QRangeModel : public QObject
 {
     Q_OBJECT
 
-    Q_PROPERTY(int minimumValue READ minimum WRITE setMinimum NOTIFY rangeChanged)
-    Q_PROPERTY(int maximumValue READ maximum WRITE setMaximum NOTIFY rangeChanged)
-    Q_PROPERTY(int singleStep READ singleStep WRITE setSingleStep)
-    Q_PROPERTY(int pageStep READ pageStep WRITE setPageStep)
-    Q_PROPERTY(int value READ value WRITE setValue NOTIFY valueChanged USER true)
-    Q_PROPERTY(int position READ position WRITE setPosition NOTIFY positionChanged)
-    Q_PROPERTY(int minimumPosition READ minimumPosition WRITE setMinimumPosition NOTIFY positionRangeChanged)
-    Q_PROPERTY(int maximumPosition READ maximumPosition WRITE setMaximumPosition NOTIFY positionRangeChanged)
+    Q_PROPERTY(qreal minimumValue READ minimum WRITE setMinimum NOTIFY rangeChanged)
+    Q_PROPERTY(qreal maximumValue READ maximum WRITE setMaximum NOTIFY rangeChanged)
+    Q_PROPERTY(qreal singleStep READ singleStep WRITE setSingleStep)
+    Q_PROPERTY(qreal pageStep READ pageStep WRITE setPageStep)
+    Q_PROPERTY(qreal value READ value WRITE setValue NOTIFY valueChanged USER true)
+    Q_PROPERTY(qreal position READ position WRITE setPosition NOTIFY positionChanged)
+    Q_PROPERTY(qreal minimumPosition READ minimumPosition WRITE setMinimumPosition NOTIFY positionRangeChanged)
+    Q_PROPERTY(qreal maximumPosition READ maximumPosition WRITE setMaximumPosition NOTIFY positionRangeChanged)
     Q_PROPERTY(bool tracking READ hasTracking WRITE setTracking)
 
 public:
@@ -53,38 +53,38 @@ public:
     void sedate();
     void awake();
 
-    void setMinimum(int);
-    int minimum() const;
+    void setMinimum(qreal);
+    qreal minimum() const;
 
-    void setMaximum(int);
-    int maximum() const;
+    void setMaximum(qreal);
+    qreal maximum() const;
 
-    void setRange(int min, int max);
+    void setRange(qreal min, qreal max);
 
-    void setSingleStep(int);
-    int singleStep() const;
+    void setSingleStep(qreal);
+    qreal singleStep() const;
 
-    void setPageStep(int);
-    int pageStep() const;
+    void setPageStep(qreal);
+    qreal pageStep() const;
 
     void setTracking(bool enable);
     bool hasTracking() const;
 
-    void setSliderPosition(int);
-    int sliderPosition() const;
+    void setSliderPosition(qreal);
+    qreal sliderPosition() const;
 
-    void setMinimumPosition(int);
-    int minimumPosition() const;
+    void setMinimumPosition(qreal);
+    qreal minimumPosition() const;
 
-    void setMaximumPosition(int);
-    int maximumPosition() const;
+    void setMaximumPosition(qreal);
+    qreal maximumPosition() const;
 
-    void setPositionRange(int min, int max);
+    void setPositionRange(qreal min, qreal max);
 
-    void setPosition(int position);
-    int position() const;
+    void setPosition(qreal position);
+    qreal position() const;
 
-    int value() const;
+    qreal value() const;
 
 public Q_SLOTS:
     void singleStepAdd();
@@ -93,16 +93,16 @@ public Q_SLOTS:
     void pageStepSub();
     void toMinimum();
     void toMaximum();
-    void setValue(int);
+    void setValue(qreal);
 
 Q_SIGNALS:
-    void sliderMoved(int position);
+    void sliderMoved(qreal position);
 
-    void valueChanged(int value);
-    void positionChanged(int position);
+    void valueChanged(qreal value);
+    void positionChanged(qreal position);
 
-    void rangeChanged(int min, int max);
-    void positionRangeChanged(int min, int max);
+    void rangeChanged(qreal min, qreal max);
+    void positionRangeChanged(qreal min, qreal max);
 
 protected:
     QRangeModel(QRangeModelPrivate &dd, QObject *parent);
